refactor(generate): isolate protocol adapters and tests
